Use nanosleep on Linux in atomic_backoff when spinning for a long time.
This allows to significantly reduce CPU system time while spinning, as
nanosleep deschedules the thread and allows other threads to progress.
Unlike sched_yield, nanosleep allows the thread to migrate to other CPUs,
as well as other threads to migrate to the current CPU, which is what
allows for more total progress.

See https://github.com/oneapi-src/oneTBB/issues/105. On the test presented
there, this patch offers to reduce CPU system time by 25-30%. On a real world
application that test is based on this patch offers ~10% improvement.