Allow runtime malloc replacement using LoadLibary
When a .dll is loaded using the LoadLibary() then the "reserved" parameter is NULL in the DllMain() function. This makes it impossible to dynamicly load the malloc_proxy dynamically at runtime because the "reserved" parameter is checked to be not null. So I removed the (I think needles) condition on that variable so that the doMallocReplacement() will be called on a LoadLibary(). This allows programmers to create applications that can "optionally" use the tbb_malloc replacement without having to re-link the application.