Fix tbb4py test sporadic fails (#913)
There are sporadic fails with tbb4py test (test.py). The problem is that some test cases rely on the fact, that submitted async work will not finish until calling get with timeout. To do there is a call for time.sleep in functor submitted. Sporadically, expected result is not happening (like OSX example in issue) so to prevent that, the blocking is used.

Signed-off-by: Isaev, Ilya <ilya.isaev@intel.com>