Fixed typo (maybe?) in CMakeLists.txt
Currently TBB attempts a find package and excludes the entire build part of the CMake setup when *either* `TBB_FIND_PACKAGE` *or* `TBB_DIR` are defined. This is an issue when using `FetchContent` with `OVERRIDE_FIND_PACKAGE` as CMake declares `TBB_DIR` to equal `${CMAKE_BINARY_DIR}/CMakeFiles/pkgRedirects/`. It also strikes me as strange that the library could skip compilation and just find itself without the request of the user. It makes more sense that both these variables must be set for the library to skip compilation.